Provide a default value for null baseUrl and path
On the RequestBuilder object, both the path and the baseUrl are optional. This signals the user that they can be omited.

These cases where not properly handled in the Ktor implementation of HttpRequestFactory. Passing in `http://www.mirego.com/` for `baseUrl` and nothing for `path` resulted in the following request url `http://www.mirego.com/null`

The Native implementation is not affected