Fix race condition when receiving the response
The BehaviourSubject we are using to return the response was freezed only after starting the request. This created a possible race condition where the response was assigned to the value of the publisher from another thread while the publisher is being frozen.

This caused various  crashes in the Kotlin/Native freezing process